get marks from pdf and set value

+++ src/main/java/com/Debuggers/MobiliteInternational/Entity/Candidacy.java
@@ -35,7 +35,7 @@ public class Candidacy implements Serializable {
     private Niveau levelFr;
     @Enumerated(EnumType.STRING)
     private Status status ;
-    private Float marks;
+    private double marks;
     private Boolean archive;
     @ManyToOne(cascade = CascadeType.ALL)
     private User user;
